requirements.txt
# absl-py==0.12.0
# appdirs==1.4.4
# argcomplete==1.12.2
# astunparse==1.6.3
# attrs==19.3.0
# aubio==0.4.9
# Automat==0.8.0
# beautifulsoup4==4.7.1
# blinker==1.4
# cachetools==4.2.1
# certifi==2019.3.9
# chardet==3.0.4
# Click==7.0
# # cloud-init==20.4
# colorama==0.4.3
# command-not-found==0.3
# configobj==5.0.6
# constantly==15.1.0
# coverage==5.5
# crcmod==1.7
# cryptography==2.8
# cssselect==1.0.3
# dbus-python==1.2.16
# distlib==0.3.2
# distro==1.4.0
# distro-info===0.23ubuntu1
# dotty-dict==1.3.0
# entrypoints==0.3
# feedfinder2==0.0.4
# feedparser==5.2.1
# filelock==3.0.12
# flake8==3.9.0
Flask==1.0.2
# Flask-Cors==3.0.7
# flatbuffers==1.12
# gast==0.3.3
# google-auth==1.28.0
# google-auth-oauthlib==0.4.3
# google-pasta==0.2.0
# goose3==3.1.7
# grpcio==1.32.0
gunicorn==20.1.0
# h5py==2.10.0
# hjson==3.0.2
# httplib2==0.14.0
# hyperlink==19.0.0
# idna==2.8
# importlib-metadata==1.5.0
# incremental==16.10.1
# itsdangerous==1.1.0
# jieba==0.42.1
# jieba3k==0.35.1
# Jinja2==2.10
# joblib==1.0.0
# jsonpatch==1.22
# jsonpointer==2.0
# jsonschema==3.2.0
# Keras==2.4.3
# keras-bert==0.86.0
# keras-embed-sim==0.8.0
# keras-layer-normalization==0.14.0
# keras-multi-head==0.27.0
# keras-pos-embd==0.11.0
# keras-position-wise-feed-forward==0.6.0
# Keras-Preprocessing==1.1.2
# keras-radam==0.15.0
# keras-self-attention==0.46.0
# keras-transformer==0.38.0
# keyring==18.0.1
# language-selector==0.1
# launchpadlib==1.10.13
# lazr.restfulclient==0.14.2
# lazr.uri==1.0.3
# lxml==4.6.2
# Markdown==3.3.4
# MarkupSafe==1.1.1
# mccabe==0.6.1
# milc==1.3.0
# more-itertools==4.2.0
# netifaces==0.10.4
newspaper3k==0.2.8
nltk==3.4
# nose2==0.10.0
# numpy==1.19.2
# oauthlib==3.0.1
# opt-einsum==3.3.0
# pandas==1.2.1
# pexpect==4.6.0
# Pillow==8.1.0
# protobuf==3.15.6
# pyasn1==0.4.2
# pyasn1-modules==0.2.1
# pycodestyle==2.7.0
# pyflakes==2.3.1
# Pygments==2.8.1
# PyGObject==3.36.0
# PyHamcrest==1.9.0
# PyJWT==1.7.1
# pymacaroons==0.13.0
# PyNaCl==1.3.0
# pyOpenSSL==19.0.0
# pyrsistent==0.15.5
# pyserial==3.4
# python-apt==2.0.0+ubuntu0.20.4.4
# python-dateutil==2.8.0
# python-debian===0.1.36ubuntu1
# pytz==2021.1
# PyYAML==5.1
# qmk==0.0.45
# regex==2020.11.13
# requests==2.21.0
# requests-file==1.4.3
# requests-oauthlib==1.2.0
# requests-unixsocket==0.2.0
# rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.0
# SecretStorage==2.3.1
# service-identity==18.1.0
# setuptools-scm==6.0.1
# simplejson==3.16.0
# singledispatch==3.4.0.3
# six==1.15.0
# sklearn==0.0
# sos==4.0
# soupsieve==1.8
# ssh-import-id==5.10
# systemd-python==234
# tensorboard==2.4.1
# tensorboard-plugin-wit==1.8.0
# tensorflow==2.4.1
# tensorflow-estimator==2.4.0
# termcolor==1.1.0
# threadpoolctl==2.1.0
# tinysegmenter==0.3
# tldextract==2.2.1
# tqdm==4.31.1
# Twisted==18.9.0
# twython==3.7.0
# typing-extensions==3.7.4.3
# ubuntu-advantage-tools==20.3
# ufw==0.36
# unattended-upgrades==0.1
# Unidecode==1.1.2
# urllib3==1.24.1
# virtualenv==20.4.7
# wadllib==1.3.3
# Werkzeug==2.0.1
# wrapt==1.12.1
# yapf==0.31.0
# zipp==1.0.0
# zope.interface==4.7.1
